Moved factory method for BatchingLedgerWriter to where it belongs.
