+++ ledger/participant-state/kvutils/src/main/scala/com/daml/ledger/participant/state/kvutils/api/BatchingLedgerWriter.scala
@@ -16,6 +16,7 @@ import com.daml.logging.LoggingContext.newLoggingContext
 import com.daml.logging.{ContextualizedLogger, LoggingContext}
 
 import scala.collection.JavaConverters._
+import scala.concurrent.duration.{Duration, MILLISECONDS}
 import scala.concurrent.{ExecutionContext, Future}
 
 /** A batching ledger writer that collects submissions into a batch and commits
@@ -76,4 +77,34 @@ class BatchingLedgerWriter(val queue: BatchingQueue, val writer: LedgerWriter)(
   }
 
   override def close(): Unit = queueHandle.close()
+}
+
+object BatchingLedgerWriter {
+  def apply(batchingLedgerWriterConfig: BatchingLedgerWriterConfig, delegate: LedgerWriter)(
+      implicit materializer: Materializer,
+      loggingContext: LoggingContext): BatchingLedgerWriter = {
+    val batchingQueue = batchingQueueFrom(batchingLedgerWriterConfig)
+    new BatchingLedgerWriter(batchingQueue, delegate)
+  }
+
+  private def batchingQueueFrom(
+      batchingLedgerWriterConfig: BatchingLedgerWriterConfig): BatchingQueue =
+    if (batchingLedgerWriterConfig.enableBatching) {
+      DefaultBatchingQueue(
+        maxQueueSize = batchingLedgerWriterConfig.maxBatchQueueSize,
+        maxBatchSizeBytes = batchingLedgerWriterConfig.maxBatchSizeBytes,
+        maxWaitDuration = batchingLedgerWriterConfig.maxBatchWaitDuration,
+        maxConcurrentCommits = batchingLedgerWriterConfig.maxBatchConcurrentCommits
+      )
+    } else {
+      batchingQueueForSerialValidation(batchingLedgerWriterConfig.maxBatchQueueSize)
+    }
+
+  private def batchingQueueForSerialValidation(maxBatchQueueSize: Int): DefaultBatchingQueue =
+    DefaultBatchingQueue(
+      maxQueueSize = maxBatchQueueSize,
+      maxBatchSizeBytes = 1,
+      maxWaitDuration = Duration(1, MILLISECONDS),
+      maxConcurrentCommits = 1
+    )
 }
