+++ ledger/participant-state/kvutils/src/main/scala/com/daml/ledger/validator/batch/BatchedSubmissionValidatorFactory.scala
@@ -3,34 +3,13 @@
 
 package com.daml.ledger.validator.batch
 
-import akka.stream.Materializer
 import com.daml.caching.Cache
 import com.daml.ledger.participant.state.kvutils.DamlKvutils.{DamlStateKey, DamlStateValue}
-import com.daml.ledger.participant.state.kvutils.api.{
-  BatchingLedgerWriter,
-  BatchingLedgerWriterConfig,
-  BatchingQueue,
-  DefaultBatchingQueue,
-  LedgerWriter
-}
 import com.daml.ledger.validator.LedgerStateOperations.{Key, Value}
-import com.daml.ledger.validator.{
-  CachingCommitStrategy,
-  CachingDamlLedgerStateReader,
-  CommitStrategy,
-  DamlLedgerStateReader,
-  DefaultStateKeySerializationStrategy,
-  LedgerStateOperations,
-  LedgerStateReader,
-  LogAppendingCommitStrategy,
-  QueryableReadSet,
-  StateKeySerializationStrategy
-}
+import com.daml.ledger.validator._
 import com.daml.lf.engine.Engine
-import com.daml.logging.LoggingContext
 import com.daml.metrics.Metrics
 
-import scala.concurrent.duration.{Duration, MILLISECONDS}
 import scala.concurrent.{ExecutionContext, Future}
 
 object BatchedSubmissionValidatorFactory {
@@ -44,36 +23,6 @@ object BatchedSubmissionValidatorFactory {
         commitParallelism = 1
       )
     }
-
-  def batchingLedgerWriterFrom(
-      batchingLedgerWriterConfig: BatchingLedgerWriterConfig,
-      delegate: LedgerWriter)(
-      implicit materializer: Materializer,
-      loggingContext: LoggingContext): BatchingLedgerWriter = {
-    val batchingQueue = batchingQueueFrom(batchingLedgerWriterConfig)
-    new BatchingLedgerWriter(batchingQueue, delegate)
-  }
-
-  private def batchingQueueFrom(
-      batchingLedgerWriterConfig: BatchingLedgerWriterConfig): BatchingQueue =
-    if (batchingLedgerWriterConfig.enableBatching) {
-      DefaultBatchingQueue(
-        maxQueueSize = batchingLedgerWriterConfig.maxBatchQueueSize,
-        maxBatchSizeBytes = batchingLedgerWriterConfig.maxBatchSizeBytes,
-        maxWaitDuration = batchingLedgerWriterConfig.maxBatchWaitDuration,
-        maxConcurrentCommits = batchingLedgerWriterConfig.maxBatchConcurrentCommits
-      )
-    } else {
-      batchingQueueForSerialValidation(batchingLedgerWriterConfig.maxBatchQueueSize)
-    }
-
-  private def batchingQueueForSerialValidation(maxBatchQueueSize: Int): DefaultBatchingQueue =
-    DefaultBatchingQueue(
-      maxQueueSize = maxBatchQueueSize,
-      maxBatchSizeBytes = 1,
-      maxWaitDuration = Duration(1, MILLISECONDS),
-      maxConcurrentCommits = 1
-    )
 
   class LedgerStateReaderAdapter[LogResult](delegate: LedgerStateOperations[LogResult])
       extends LedgerStateReader {
